TECH: Extract lens data to hook
